moved location JS/handlebar content to metric JS/handlebar files, added new metric-routes file
