+++ views/layouts/main.handlebars
@@ -30,7 +30,7 @@
 				<nav class="sidebar col-12 col-sm-2 col-xl-1">
 					<a class="nav-choice" href="/">Recent</a>
 					{{! the 'Recent' nav button will act as 'homepage' }}
-					<a class="nav-choice" href="/locations">Locations</a>
+					<a class="nav-choice" href="/metrics">Locations</a>
 					<a class="nav-choice" href="/users">Users</a>
 					<a class="nav-choice" href="/popular/locations">Popular</a>
 					<a class="nav-choice" href="/photos">Photos</a>
